pallet_address_associations::pallet::dispatchables

Function associate_address

pub fn associate_address<T: Config>(
    partnerchain_address: T::PartnerChainAddress,
    signature: StakeKeySignature,
    stake_public_key: StakePublicKey,
)

Extrinsic creating a new address association.

signature is expected to be a signature of the Cardano private key corresponding to stake_public_key of AddressAssociationSignedMessage created using the associated public key, address and the genesis UTXO of the particular Partner Chain it is being submitted to.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::associate_address.